Added bitmaps to main model.
Instead of maintaining two lists, a new object `SourceImage` is used that provides all information about an image - filename, bitmap etc.

With this change we no longer need to access main form from image selection form.
